+++ src-ui/js/ui/Listener.js
@@ -117,8 +117,9 @@ ui.listener =
 
 		ui.setdisplay('keypopup');
 		ui.setdisplay('bgcolor');
-		ui.setdisplay('passallcell');
-		ui.setdisplay('aquarium_regions');
+		for(var key in ui.puzzle.config.getVariants()){
+			ui.setdisplay(key);
+		}
 		ui.keypopup.display();
 	},